+++ tests/test_functions.py
@@ -1,10 +1,8 @@
 import sys
 import io
-import os
 import unittest
 from unittest.mock import patch, DEFAULT
 import requests
-import filecmp
 from operator import attrgetter
 
 # add parent directory to imports list
@@ -18,38 +16,15 @@ sys. path.insert(0,'./xmls')
 
 # local import
 
-from constants import ESUMMARY_URL, ESEARCH_URL
-from functions import dispatch, download, esearchquery
+from constants import ESEARCH_URL
+from functions import download, esearchquery
 from data import data, esummary_response  #METAZOA, FUNGI, PLANTAE, WRONG_LINEAGE, CUSTOM_TAXONOMY
 METAZOA, FUNGI, PLANTAE, WRONG_LINEAGE, CUSTOM_TAXONOMY = attrgetter("METAZOA", "FUNGI", "PLANTAE", "WRONG_LINEAGE", "CUSTOM_TAXONOMY")(data)
 mocked_response_xms , expected_output = attrgetter("mocked_response_xms", "expected_output")(esummary_response)
 
 class testsFunctions(unittest.TestCase):
 
-    def test_dispatch(self):
-    
-        ##kingdoms
-        self.assertEqual(dispatch(METAZOA, 1), "METAZOA")
-        self.assertEqual(dispatch(PLANTAE, 1), "PLANTAE")
-        self.assertEqual(dispatch(FUNGI, 1), "FUNGI")
-        self.assertEqual(dispatch(WRONG_LINEAGE, 1), "OTHERS")
-        ##no option selected
-        self.assertEqual(dispatch(METAZOA, 2), "sequences")
-        self.assertEqual(dispatch(FUNGI, 2), "sequences")
-        ##phylums
-        self.assertEqual(dispatch(WRONG_LINEAGE, 0), "OTHERS")
-        self.assertEqual(dispatch(PLANTAE, 0), 'Chlorophyta')
-        self.assertEqual(dispatch(FUNGI, 0),'Ascomycota')
-        self.assertEqual(dispatch(METAZOA,0), 'Chordata')
-        ##list of taxonomic levels
-        self.assertEqual(dispatch(METAZOA, CUSTOM_TAXONOMY), 'Primates')
-        self.assertEqual(dispatch(METAZOA, WRONG_LINEAGE), 'OTHERS')
-        ## n rank higher than specie
-        self.assertEqual(dispatch(METAZOA, 3), 'Homo.')
-        self.assertEqual(dispatch(PLANTAE, 5), 'Chlorellaceae')
-        self.assertEqual(dispatch(PLANTAE, 50), 'OTHERS')
-    
-
+   
     @unittest.mock.patch('sys.stdout', new_callable=io.StringIO)
     def test_download(self, mock_stdout): 
         ##parameters address
